fix: Double monetary types to BigDecimal

+++ src/main/java/com/apollo/main/dto/request/FuncionarioRequestDTO.java
@@ -3,11 +3,12 @@ package com.apollo.main.dto.request;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
 public class FuncionarioRequestDTO extends PessoaRequestDTO{
     private LocalDateTime dataAdmissao;
-    private Double salario;
+    private BigDecimal salario;
 }
